src/claudia/requirements.txt
behave==1.2.6
click==8.1.3
PyHamcrest==2.0.4
setuptools==67.7.2
xrpl-py==2.4.0
requests
inquirer==3.1.3
streamlit
watchdog
streamlit-option-menu
boto3
paramiko
